+++ hassio-google-drive-backup/backup/module.py
@@ -9,6 +9,7 @@ from typing import List
 from backup.config import Config, Startable, Setting
 from backup.drive import DriveSource
 from backup.ha import HaSource, HaUpdater, AddonStopper
+from backup.ha.integrationws import IntegrationWebSocketServer
 from backup.model import BackupDestination, BackupSource, Scyncer
 from backup.util import Resolver
 from backup.model import Coordinator, Precache, DestinationPrecache
@@ -52,9 +53,10 @@ class BaseModule(Module):
     @multiprovider
     @singleton
     def getStartables(self, debug_server: DebugServer, ha_updater: HaUpdater, debugger: DebugWorker, ha_source: HaSource,
-                      server: UiServer, restarter: Restarter, syncer: Scyncer, watcher: Watcher, stopper: AddonStopper, precache: Precache) -> List[Startable]:
+                      server: UiServer, restarter: Restarter, syncer: Scyncer, watcher: Watcher, stopper: AddonStopper, 
+                      precache: Precache, integration_ws: IntegrationWebSocketServer) -> List[Startable]:
         # Order here matters, since its the order in which components of the addon are initialized.
-        return [debug_server, ha_updater, debugger, ha_source, server, restarter, syncer, watcher, stopper, precache]
+        return [debug_server, integration_ws, ha_updater, debugger, ha_source, server, restarter, syncer, watcher, stopper, precache]
 
     @provider
     @singleton