+++ hassio-google-drive-backup/backup/static/js/scripts.js
@@ -131,14 +131,6 @@ function exposeServer(expose) {
   }, null, "Saving setting...");
 }
 
-function callBackupSnapshot(doSwitch) {
-  var url = "callbackupsnapshot?switch=" + doSwitch;
-  postJson(url, {}, function (data) {
-    $('#backup_upgrade_card').fadeOut(500);
-    refreshstats();
-  }, null, "Saving...");
-}
-
 function dismissIgnoredBackupWarning(doSwitch) {
   var url = "ignoredbackupswitch?switch=" + doSwitch;
   postJson(url, {}, function (data) {
@@ -749,8 +741,6 @@ function processStatusUpdate(data) {
     question_card = "ingress_upgrade_card";
   } else if (data.ask_error_reports && !found) {
     question_card = "error_reports_card";
-  } else if (data.warn_backup_upgrade && !found) {
-    question_card = "backup_upgrade_card";
   } else if (data.warn_upgrade_backups) {
     question_card = "warn_upgrade_backups_card";
   } else if (data.warn_oob_oauth) {